#460 Funzionamento anche senza Redis

+++ core/class/Entita.class.php
@@ -132,6 +132,7 @@ class Entita {
      */
     protected static function _invalidaCacheQuery() {
         global $cache, $conf;
+        if ( !$cache ) { return false; }
         foreach ( $cache->keys($conf['db_hash'] . static::$_t . ':query:*') as $chiave ) {
             $cache->delete($chiave);
         }
